added redis layer

+++ src/main/java/com/niket/onlineofflineindicator/records/UserActivity.java
@@ -1,4 +1,4 @@
 package com.niket.onlineofflineindicator.records;
 
-public record UserActivity(int customerId, int lastSeen, boolean isActive) {
+public record UserActivity(int customerId, long lastSeen, boolean isActive) {
 }
